Remove redundancies in 'test_*latency'
The three scripts `test_localIPP_latency.py`,
`test_localThreads_latency.py`, and `test_midway_latency.py` are all doing
the same thing, so I've collapsed them into a single script. You can
control the config by running, for example: `python latency.py --config
configs/midway_ipp.py`. Because these are not tests meant to be run in
pytest (they do not assert anything) but is rather a diagnostic script, I
am moving it out of the pytest directory.
